Fixing blueprint export problems with undisguised hostnames
This patch addresses AMBARI-7368.

The Blueprint configuration processor was not properly
  masking the hostnames in some properties after a
  blueprint export from a running cluster.

Most of these properties were related to HA or
  Kerberos, which are not supported by the Blueprints
  processor directly.

This patch addresses this problem by registering
  PropertyUpdater instances for each of the properties
  that includes hostname information, but is not currently
  masked during a blueprint export.

This patch implements support for tagging the HA Namenode configuration
  properties as candidates for hostname masking.  The HA properties
  are dynamic with respect to naming, since the properties include
  the nameserivces and name nodes defined during HA setup in the
  Ambari UI.  This patch implements a method to inspect the HA
  configuration, if present, and register the HA properties for
  each NameService and NameNode in order to remove the hostname
  information during the exporting process.

This patch also implements support for the Namenode HA configuration
  properties to be properly handled with hostname resolution during
  cluster creation.

This patch implements several new unit test cases to verify this change, and also updates some existing test cases to verify this change as well.