Adds Blueprint export handler for Hive Zookeeper quorum configuration
This patch addresses bug AMBARI-7592.

In the case of a multi-node cluster deployment that contains Hive, the
  following property:

"hive.zookeeper.quorum"

was not being properly handled by the BlueprintConfigurationProcessor.

This meant that this property would include hostname information
  in an exported Blueprint.  This is incorrect, since the Blueprint
  should not contain direct references to hostnames.

This patch fixes this bug by registering a PropertyUpdater
  instance for "hive.zookeeper.quorum" in hive-site.xml.  The
  PropertyUpdater is responsible for stripping out the
  hostnames, and adding in the expected tokens and host
  group names.

This patch also adds new test assertions to an existing
  unit test in order to verify this fix.