Fix for agent setup failing on localhost
When installing the agent on the same machine (for testing or learning purposes) using "localhost.whatever" as FQDN, the agent configuration is messed up by `configureAgent`. As the obvious purpose there is to replace the default "hostname=localhost" with whatever the user is provided, it makes sense to make the match more exact.