AMBARI-25341 SmartSense API call fails with Unsupported Media Type
BUG-121188

The AMBARI-9016 is causing regression in SmartSense by inappropriately changing the Content-Type
to "text/plain" for every request with "application/json" Content-Type. This is a proper
workaround for the Ambari API but not for SmartSense.

To fix this the ContentTypeOverrideFilter omits every "/views/SMARTSENSE/.*" requests, thus not changing
its Content-Type.

Change-Id: Iffbe42e50ba85bc24e44ee49149fd435e4b4af66