AMBARI-25635 Clear Cluster and METRIC_AGGREGATORS MBeans upon shutdown (#3304) (payert via dgrinenko)
Refactor HelicController inner class to LiveInstanceTracker.

The HelicController was derived from GeneralHelixController class. This caused problem becasue
practically two Controllers were in use due to thus bad pactice.
The HelicController class is not intended to be a real Controller, only a tracker of the live nodes
what is basically just a LiveInstanceChangeListener implementation.