Fix the widget.json for the Host memory used %
'free' memory is the memory available after the cached memory getting reduced from total memory but the 'available' memory is the total available memory including the cached memory as that can be reclaimed anytime. So mem_free is nothing but mem_available. So in the host memory used heat map we should just do (mem_total-mem_free)/mem_total to get the used memory fraction.